+++ CHANGELOG.md
@@ -36,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - forwardconnector
   - countconnector
 
+- feat(sumologicexporter): use otlp url suffixes [#1015]
+
 [#969]: https://github.com/SumoLogic/sumologic-otel-collector/pull/969
 [#975]: https://github.com/SumoLogic/sumologic-otel-collector/pull/975
 [#980]: https://github.com/SumoLogic/sumologic-otel-collector/pull/980
@@ -43,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#1011]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1011
 [#1013]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1013
 [#1016]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1016
+[#1015]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1015
 [unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.71.0-sumo-0...main
 
 ## [v0.71.0-sumo-0]